Moved Graphics Management Tools to Common
