+++ src/arcade/common/graphicsManager.py
