+++ src/arcade/common/spritesheet.py
